✨ `upgrade device` now supports multiple devices.

+++ centralcli/clitree/upgrade.py
@@ -8,8 +8,9 @@ from typing import TYPE_CHECKING
 import typer
 from rich.markup import escape
 
-from centralcli import common, render
+from centralcli import common, render, utils
 from centralcli.cache import api
+from centralcli.client import BatchRequest
 from centralcli.constants import AllDevTypes, iden_meta, lib_to_gen_plural  # noqa
 from centralcli.objects import DateTime
 
@@ -21,7 +22,7 @@ app = typer.Typer()
 
 @app.command()
 def device(
-    device: str = common.arguments.device,
+    devices: list[str] = common.arguments.devices,
     version: str = common.arguments.version,
     at: datetime = common.options.at,
     in_: str = common.options.in_,
@@ -31,48 +32,57 @@ def device(
     default: bool = common.options.default,
     workspace: str = common.options.workspace,
 ) -> None:
-    """Upgrade [dim italic](or Downgrade)[/] firmware on a device
+    """Upgrade [dim italic](or Downgrade)[/] firmware on device(s).
     """
-    forced = None
-    dev = common.cache.get_dev_identifier(device, conductor_only=True)
-    if dev.generic_type == "ap":
-        reboot = True
-        if version and "beta" in version:  # beta for APs always looks like this "10.7.1.0-10.7.1.0-beta_91138"
-            needless_prefix = version.split("-")[0]
-            if not version.count(needless_prefix) == 2:
-                version = f"{needless_prefix}-{version}"
-    elif dev.type == "gw":
-        forced = True
+    devs = [common.cache.get_dev_identifier(dev, conductor_only=True) for dev in devices]
+    dev_types = list(set([dev.type for dev in devs]))
+    if len([t for t in dev_types if t not in ["ap", "gw"]]) > 1:
+        common.exit(f"Specifying multiple devices of different types ({utils.summarize_list(dev_types, pad=0, sep=', ')}) does not make sense.  All devices should be compatible with the same software/version.")
+
+    batch_reqs = []
+    ap_version = None
+    for dev in devs:
+        if dev.generic_type == "ap":
+            ap_version = version
+            if version and "beta" in version:  # beta for APs always looks like this "10.7.1.0-10.7.1.0-beta_91138"
+                needless_prefix = version.split("-")[0]
+                if not version.count(needless_prefix) == 2:
+                    ap_version = f"{needless_prefix}-{version}"
+            if not dev.swack_id:
+                render.econsole.print(f"\n[cyan]{dev.name}[/] lacks a swarm_id, may not be populated yet if it was recently added.")
+                if yes > 1 or render.confirm(prompt="\nRefresh cache now to check if it's populated?"):
+                    api.session.request(common.cache.refresh_dev_db, dev_type="ap")
+                    dev = common.cache.get_dev_identifier(dev.serial, dev_type="ap")
+                if not dev.swack_id:
+                    common.exit(f"Unable to perform Upgrade on {dev.summary_text}.  [cyan]swarm_id[/] is required for APs and the API is not returning a value for it yet.")
+            batch_reqs += [BatchRequest(api.firmware.upgrade_firmware, scheduled_at=at, swarm_id=dev.swack_id, firmware_version=ap_version)]
+        else:
+            batch_reqs += [BatchRequest(api.firmware.upgrade_firmware, scheduled_at=at, serial=dev.serial, firmware_version=version, reboot=reboot, forced=None if not dev.type == "gw" else True)]
+
     at = None if not at else round(at.timestamp())
     if in_:
         at = common.delta_to_start(in_, past=False).int_timestamp
 
-
-    ver_msg = "Recommended version" if not version else version
-    ver_msg = f'Upgrade [cyan]{dev.name}[/] to [green3]{ver_msg}[/]'
+    ver_msg = f"Upgrad{'ing' if yes else 'e'}"
+    ver_msg = f"{ver_msg} [cyan]{dev.name}[/]" if len(devs) == 1 else f"{ver_msg} the following devices"
+    ver_msg = f"{ver_msg} to the recommended version" if not version else f"{ver_msg} to [green3]{version}[/]"
     if at:
         dt = DateTime(at)
         ver_msg = f'{ver_msg} [italic]at {dt}[/]'
 
-    ver_msg = f"{ver_msg} and :recycle:  reboot" if reboot else f"{ver_msg} ('-R' not specified, [italic bright_red]device will not be rebooted[/])"
+    if len(dev_types) > 1 and not reboot:  # This can only happen with ap and gw in the list of devices
+        ver_msg = f"{ver_msg}.  :recycle:  APs will reboot [dim italic]always the case for APs[/] Gateways will not as '-R' not specified. [italic bright_red]Gateways will not be rebooted[/]!!"
+    else:
+        ver_msg = f"{ver_msg} and :recycle:  reboot" if reboot else f"{ver_msg} ('-R' not specified, [italic bright_red]device will not be rebooted[/])"
+
+    if len(devs) > 1:
+        ver_msg = f"{ver_msg}{utils.summarize_list([dev.rich_help_text for dev in devs], max=9)}"
 
     render.econsole.print(ver_msg)
     render.confirm(yes)  # aborts here if they don't confirm
+    batch_resp = api.session.batch_request(batch_reqs)
 
-    if dev.type == "ap":  # TODO need to validate this is the same behavior for 8.x IAP.
-        if not dev.swack_id:
-            render.econsole.print(f"\n[cyan]{dev.name}[/] lacks a swarm_id, may not be populated yet if it was recently added.")
-            if yes > 1 or render.confirm(prompt="\nRefresh cache now to check if it's populated"):
-                api.session.request(common.cache.refresh_dev_db, dev_type="ap")
-                dev = common.cache.get_dev_identifier(dev.serial, dev_type="ap")
-
-        if dev.swack_id:
-            resp = api.session.request(api.firmware.upgrade_firmware, scheduled_at=at, swarm_id=dev.swack_id, firmware_version=version, reboot=reboot)
-        else:
-            common.exit(f"Unable to perform Upgrade on {dev.summary_text}.  [cyan]swarm_id[/] is required for APs and the API is not returning a value for it yet.")
-    else:
-        resp = api.session.request(api.firmware.upgrade_firmware, scheduled_at=at, serial=dev.serial, firmware_version=version, reboot=reboot, forced=forced)
-    render.display_results(resp, tablefmt="action")
+    render.display_results(batch_resp, tablefmt="action")
 
 
 @app.command()
